% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{fold_change}
\alias{fold_change}
\title{Fold change}
\usage{
fold_change(object, group, assay.type = NULL)
}
\arguments{
\item{object}{a \code{
\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}
object}

\item{group}{character, name of the group column}

\item{assay.type}{character, assay to be used in case of multiple assays}
}
\value{
A data frame with fold changes for each feature.
}
\description{
Computes fold change between each group for each feature.
}
\examples{
data(toy_notame_set, package = "notame")
# Between groups
fc <- fold_change(toy_notame_set, group = "Group")
# Between time points
fc <- fold_change(toy_notame_set, group = "Time")

}
