% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate.R
\name{importance_rf}
\alias{importance_rf}
\title{Feature importance in random forest}
\usage{
importance_rf(rf)
}
\arguments{
\item{rf}{An object of class randomForest}
}
\value{
A data frame of feature importance.
}
\description{
Extracts feature importance in random forest in a nice format.
}
\examples{
data(example_set, package = "notame")
rf <- fit_rf(example_set, y = "Group")
rf
importance_rf(rf)

}
\seealso{
\code{\link[randomForest]{randomForest}}, \code{\link{fit_rf}}
}
